Move Sync Service to its own namespace
- Move Consumer to Sync service namespace
- Invalidate indexer now delegates publishing to queue or syncing products to Nosto
